Phase-2/backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
sqlmodel>=0.0.14
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
pytest>=7.0.0
ruff>=0.1.0
requests>=2.31.0
pyjwt>=2.8.0
pydantic-settings>=2.0.0
